splitting bar charts logic into components
